requirements.txt
gunicorn
Django
django-rest-framework
django-cors-headers
markdown
drf-spectacular
configparser
psycopg2
requests
bs4
